将ECommon合并为一个类库
