+++ src/ECommon/Configurations/ConfigurationExtensions.cs
@@ -1,0 +1,63 @@
+﻿using Autofac;
+using ECommon.Autofac;
+using ECommon.Components;
+using ECommon.JsonNet;
+using ECommon.Log4Net;
+using ECommon.Logging;
+using ECommon.ProtocolBuf;
+using ECommon.Serializing;
+
+namespace ECommon.Configurations
+{
+    /// <summary>ENode configuration class Autofac extensions.
+    /// </summary>
+    public static class ConfigurationExtensions
+    {
+        /// <summary>Use Autofac as the object container.
+        /// </summary>
+        /// <returns></returns>
+        public static Configuration UseAutofac(this Configuration configuration)
+        {
+            return UseAutofac(configuration, new ContainerBuilder());
+        }
+        /// <summary>Use Autofac as the object container.
+        /// </summary>
+        /// <returns></returns>
+        public static Configuration UseAutofac(this Configuration configuration, ContainerBuilder containerBuilder)
+        {
+            ObjectContainer.SetContainer(new AutofacObjectContainer(containerBuilder));
+            return configuration;
+        }
+        /// <summary>Use Json.Net as the json serializer.
+        /// </summary>
+        /// <returns></returns>
+        public static Configuration UseJsonNet(this Configuration configuration)
+        {
+            configuration.SetDefault<IJsonSerializer, NewtonsoftJsonSerializer>(new NewtonsoftJsonSerializer());
+            return configuration;
+        }
+        /// <summary>Use Log4Net as the logger.
+        /// </summary>
+        /// <returns></returns>
+        public static Configuration UseLog4Net(this Configuration configuration)
+        {
+            return UseLog4Net(configuration, "log4net.config");
+        }
+        /// <summary>Use Log4Net as the logger.
+        /// </summary>
+        /// <returns></returns>
+        public static Configuration UseLog4Net(this Configuration configuration, string configFile)
+        {
+            configuration.SetDefault<ILoggerFactory, Log4NetLoggerFactory>(new Log4NetLoggerFactory(configFile));
+            return configuration;
+        }
+        /// <summary>Use ProtocolBufSerializer as the binary serializer.
+        /// </summary>
+        /// <returns></returns>
+        public static Configuration UseProtoBufSerializer(this Configuration configuration)
+        {
+            configuration.SetDefault<IBinarySerializer, ProtocolBufSerializer>(new ProtocolBufSerializer());
+            return configuration;
+        }
+    }
+}
