--- src/Extensions/ECommon.Autofac/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Autofac" version="3.5.2" targetFramework="net45" />
-</packages>
