--- src/Extensions/ECommon.ProtocolBuf/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="protobuf-net" version="2.0.0.668" targetFramework="net45" />
-</packages>
