+++ src/RemotingPerformanceTest.Client/Program.cs
@@ -4,9 +4,8 @@ using System.Net;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using ECommon.Autofac;
 using ECommon.Components;
-using ECommon.Log4Net;
+using ECommon.Configurations;
 using ECommon.Logging;
 using ECommon.Remoting;
 using ECommon.Scheduling;
