--- src/RemotingPerformanceTest.Client/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Autofac" version="3.5.2" targetFramework="net45" />
-  <package id="log4net" version="2.0.3" targetFramework="net45" />
-</packages>
